+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - ReferenceProvider provides references for symbol macros.
 - DocumentHighlightProvider provides document highlighting for messages, tasks, actions and symbol macros.
+- Provide diagnostics for undefined `%symbol` macros.
 
 ### Improved
 - Use stored quest instances for the following providers: